Add support for overriding file based data with results from lookups.
This commit introduces the ability override variables stored in a
(YAML-)file.

The variable names stored in the file are the only pieces of information
we have when formulating a jerakia query. This implies that the
variable names in the file must follow a naming standard that let us
infer the namespace, key and metadata in order to let us do a precise
jerakia query. Right now the naming standard i severly limited (and
should be improved!): If a variable in the file is named 'foo_bar', the
query namespace is 'foo' and the key is 'bar'. Currently, query metadata
is not handled.

A test case for overriding data is also provided.